rename: vk: pool and descriptor set
