+++ Source/API/Vulkan/DescriptorPoolVk.h
@@ -1,101 +1,101 @@
-#pragma once
-
-namespace buma3d
-{
-
-class B3D_API DescriptorPoolVk : public IDeviceChildVk<IDescriptorPool>, public util::details::NEW_DELETE_OVERRIDE
-{
-protected:
-    B3D_APIENTRY DescriptorPoolVk();
-    DescriptorPoolVk(const DescriptorPoolVk&) = delete;
-    B3D_APIENTRY ~DescriptorPoolVk();
-
-private:
-    BMRESULT B3D_APIENTRY Init(DeviceVk* _device, const DESCRIPTOR_POOL_DESC& _desc);
-    void B3D_APIENTRY CopyDesc(const DESCRIPTOR_POOL_DESC& _desc);
-    void B3D_APIENTRY GetNativePoolSizes(util::DyArray<VkDescriptorPoolSize>* _sises);
-    BMRESULT B3D_APIENTRY CreateVkDescriptorPool();
-    void B3D_APIENTRY Uninit();
-
-public:
-    static BMRESULT
-        B3D_APIENTRY Create(DeviceVk* _device, const DESCRIPTOR_POOL_DESC& _desc, DescriptorPoolVk** _dst);
-
-    void
-        B3D_APIENTRY AddRef() override;
-
-    uint32_t
-        B3D_APIENTRY Release() override;
-
-    uint32_t
-        B3D_APIENTRY GetRefCount() const override;
-
-    const char*
-        B3D_APIENTRY GetName() const override;
-
-    BMRESULT
-        B3D_APIENTRY SetName(const char* _name) override;
-
-    IDevice*
-        B3D_APIENTRY GetDevice() const override;
-
-    const VkAllocationCallbacks*
-        B3D_APIENTRY GetVkAllocationCallbacks() const override;
-
-    const InstancePFN&
-        B3D_APIENTRY GetIsntancePFN() const override;
-
-    const DevicePFN&
-        B3D_APIENTRY GetDevicePFN() const override;
-
-    const DESCRIPTOR_POOL_DESC&
-        B3D_APIENTRY GetDesc() const override;
-
-    uint32_t
-        B3D_APIENTRY GetCurrentAllocationCount() override;
-
-    void
-        B3D_APIENTRY ResetPoolAndInvalidateAllocatedSets() override;
-
-    BMRESULT
-        B3D_APIENTRY AllocateDescriptorSet(IRootSignature* _root_signature, IDescriptorSet** _dst) override;
-
-    VkDescriptorPool
-        B3D_APIENTRY GetVkDescriptorPool() const;
-
-    uint64_t
-        B3D_APIENTRY GetResetID();
-
-    // DescriptorSetVk内部で使用します。
-    BMRESULT
-        B3D_APIENTRY AllocateDescriptors(DescriptorSetVk* _set);
-
-    // DescriptorSetVk内部で使用します。
-    void
-        B3D_APIENTRY FreeDescriptors(DescriptorSetVk* _set);
-
-private:
-    struct DESC_DATA
-    {
-        util::DyArray<DESCRIPTOR_POOL_SIZE> pool_sizes;
-    };
-
-private:
-    std::atomic_uint32_t                                            ref_count;
-    util::UniquePtr<util::NameableObjStr>                           name;
-    DeviceVk*                                                       device;
-    DESCRIPTOR_POOL_DESC                                            desc;
-    DESC_DATA                                                       desc_data;
-    util::StArray<uint32_t, DESCRIPTOR_TYPE_NUM_TYPES>              pool_remains;
-    std::mutex                                                      allocation_mutex;
-    std::atomic_uint32_t                                            allocation_count;
-    std::atomic_uint64_t                                            reset_id;
-    VkDevice                                                        vkdevice;
-    const InstancePFN*                                              inspfn;
-    const DevicePFN*                                                devpfn;
-    VkDescriptorPool                                                descriptor_pool;
-
-};
-
-
-}// namespace buma3d
+//#pragma once
+//
+//namespace buma3d
+//{
+//
+//class B3D_API DescriptorPoolVk : public IDeviceChildVk<IDescriptorPool>, public util::details::NEW_DELETE_OVERRIDE
+//{
+//protected:
+//    B3D_APIENTRY DescriptorPoolVk();
+//    DescriptorPoolVk(const DescriptorPoolVk&) = delete;
+//    B3D_APIENTRY ~DescriptorPoolVk();
+//
+//private:
+//    BMRESULT B3D_APIENTRY Init(DeviceVk* _device, const DESCRIPTOR_POOL_DESC& _desc);
+//    void B3D_APIENTRY CopyDesc(const DESCRIPTOR_POOL_DESC& _desc);
+//    void B3D_APIENTRY GetNativePoolSizes(util::DyArray<VkDescriptorPoolSize>* _sises);
+//    BMRESULT B3D_APIENTRY CreateVkDescriptorPool();
+//    void B3D_APIENTRY Uninit();
+//
+//public:
+//    static BMRESULT
+//        B3D_APIENTRY Create(DeviceVk* _device, const DESCRIPTOR_POOL_DESC& _desc, DescriptorPoolVk** _dst);
+//
+//    void
+//        B3D_APIENTRY AddRef() override;
+//
+//    uint32_t
+//        B3D_APIENTRY Release() override;
+//
+//    uint32_t
+//        B3D_APIENTRY GetRefCount() const override;
+//
+//    const char*
+//        B3D_APIENTRY GetName() const override;
+//
+//    BMRESULT
+//        B3D_APIENTRY SetName(const char* _name) override;
+//
+//    IDevice*
+//        B3D_APIENTRY GetDevice() const override;
+//
+//    const VkAllocationCallbacks*
+//        B3D_APIENTRY GetVkAllocationCallbacks() const override;
+//
+//    const InstancePFN&
+//        B3D_APIENTRY GetIsntancePFN() const override;
+//
+//    const DevicePFN&
+//        B3D_APIENTRY GetDevicePFN() const override;
+//
+//    const DESCRIPTOR_POOL_DESC&
+//        B3D_APIENTRY GetDesc() const override;
+//
+//    uint32_t
+//        B3D_APIENTRY GetCurrentAllocationCount() override;
+//
+//    void
+//        B3D_APIENTRY ResetPoolAndInvalidateAllocatedSets() override;
+//
+//    BMRESULT
+//        B3D_APIENTRY AllocateDescriptorSet(IRootSignature* _root_signature, IDescriptorSet** _dst) override;
+//
+//    VkDescriptorPool
+//        B3D_APIENTRY GetVkDescriptorPool() const;
+//
+//    uint64_t
+//        B3D_APIENTRY GetResetID();
+//
+//    // DescriptorSetVk内部で使用します。
+//    BMRESULT
+//        B3D_APIENTRY AllocateDescriptors(DescriptorSetVk* _set);
+//
+//    // DescriptorSetVk内部で使用します。
+//    void
+//        B3D_APIENTRY FreeDescriptors(DescriptorSetVk* _set);
+//
+//private:
+//    struct DESC_DATA
+//    {
+//        util::DyArray<DESCRIPTOR_POOL_SIZE> pool_sizes;
+//    };
+//
+//private:
+//    std::atomic_uint32_t                                            ref_count;
+//    util::UniquePtr<util::NameableObjStr>                           name;
+//    DeviceVk*                                                       device;
+//    DESCRIPTOR_POOL_DESC                                            desc;
+//    DESC_DATA                                                       desc_data;
+//    util::StArray<uint32_t, DESCRIPTOR_TYPE_NUM_TYPES>              pool_remains;
+//    std::mutex                                                      allocation_mutex;
+//    std::atomic_uint32_t                                            allocation_count;
+//    std::atomic_uint64_t                                            reset_id;
+//    VkDevice                                                        vkdevice;
+//    const InstancePFN*                                              inspfn;
+//    const DevicePFN*                                                devpfn;
+//    VkDescriptorPool                                                descriptor_pool;
+//
+//};
+//
+//
+//}// namespace buma3d
