+++ Source/API/Vulkan/RenderTargetViewVk.cpp
@@ -347,7 +347,7 @@ B3D_APIENTRY RenderTargetViewVk::GetVkImageSubresourceRange() const
 BMRESULT
 B3D_APIENTRY RenderTargetViewVk::AddDescriptorWriteRange(void* _dst, uint32_t _array_index) const
 {
-    auto&& dst = *RCAST<DescriptorSetVk::UPDATE_DESCRIPTOR_RANGE_BUFFER*>(_dst);
+    auto&& dst = *RCAST<DescriptorSet0Vk::UPDATE_DESCRIPTOR_RANGE_BUFFER*>(_dst);
     if (!dst.image_infos_data)
         return BMRESULT_FAILED;
     auto&& info = dst.image_infos_data[_array_index];
