+++ Source/API/Vulkan/SamplerViewVk.cpp
@@ -209,7 +209,7 @@ B3D_APIENTRY SamplerViewVk::GetVkSampler() const
 BMRESULT
 B3D_APIENTRY SamplerViewVk::AddDescriptorWriteRange(void* _dst, uint32_t _array_index) const
 {
-    auto&& dst = *RCAST<DescriptorSetVk::UPDATE_DESCRIPTOR_RANGE_BUFFER*>(_dst);
+    auto&& dst = *RCAST<DescriptorSet0Vk::UPDATE_DESCRIPTOR_RANGE_BUFFER*>(_dst);
     if (!dst.image_infos_data)
         return BMRESULT_FAILED;
     auto&& info = dst.image_infos_data[_array_index];
